pyproject.toml
[project]
name = "ai-playground"
version = "0.1.0"
description = "AI model testing playground"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv",
    "openai",
    "anthropic",
    "ipython",
    "requests>=2.32.3",
    "gradio>=5.30.0",
    "pypdf>=5.5.0",
    "openai-agents>=0.0.15",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."] 
